Keep the receive pane fast in long, busy sessions
The report was that the application gets heavy once a lot of data has
gone through the console. Measuring it (scripts/bench_terminal.py, new)
showed the cost is not a leak and not superlinear growth — it is simply
that drawing every byte at a fixed 30 fps is a policy that stops working
once a frame costs more than a frame is worth: on a machine ~6x slower
than this one, 2 Mbit/s of ASCII spent 51% of the UI thread inside the
text widget, and the worst frame was 34 ms.

Capture and display are now decoupled. Three limits, all of which give
up pixels and never bytes — the buffer, the exports and the counters
stay complete:

* FrameGovernor: every update times itself (insertion plus the repaint
  it triggers) and the pane asks the service for larger batches, less
  often, until rendering fits in ~30% of wall-clock time. Fewer, larger
  updates cost less in total than many small ones, so the log keeps up
  while the rest of the UI stays free. 30 fps down to 5 fps, automatic.
* RenderBudget: bytes-per-millisecond is measured, and anything beyond
  one frame's allowance is skipped — now also inside a single oversized
  chunk, which the old flood guard could not split, so one 64 KB burst
  could stall the very frame the guard existed to protect. An amber
  'display limited - N MB not shown' badge says so while it lasts.
* Nothing is rendered while the window is minimised, and an idle port
  polls at 150 ms instead of 33 ms.

Plus the hot paths themselves: format_hex is ~32x faster and the hex
dump ~7x (bytes.hex/translate instead of per-byte Python), timestamps
use one str.replace instead of a split/join loop, and neighbouring
same-direction chunks become a single insertion. Hex output is asserted
byte-for-byte identical to the previous implementation.

On the simulated slow machine the duty cycle drops 51% -> 37% and frame
cost stays flat across the session (x1.15); at 115200 baud the pane now
uses under 10% of the UI thread. 387 tests pass (29 new).

+++ scripts/promo/build_narration.py
@@ -51,7 +51,7 @@ def main() -> int:
         path = args.vo / f"{name}.wav"
         inputs += ["-i", str(path)]
         length = duration(args.ffprobe, path) / args.tempo
-        delay_ms = int(round(cursor * 1000))
+        delay_ms = round(cursor * 1000)
         filters.append(
             f"[{index}:a]atempo={args.tempo},"
             f"adelay={delay_ms}|{delay_ms},"
